server/requirements.txt
asgiref==3.5.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.8"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.8"
certifi==2021.10.8; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
cffi==1.15.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3"
coreapi==2.3.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
coreschema==0.0.4; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
cryptography==36.0.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
defusedxml==0.7.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
dj-database-url==0.5.0
django-cors-headers==3.11.0; python_version >= "3.7"
django-environ==0.8.1; python_version >= "3.4" and python_version < "4"
django-extensions==3.1.5; python_version >= "3.6"
django-on-heroku==1.1.2
django-templated-mail==1.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
django==4.0.3; python_version >= "3.8"
djangorestframework-simplejwt==4.8.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
djoser==2.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.5"
itypes==1.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
jinja2==3.0.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
markupsafe==2.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
oauthlib==3.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
psycopg2-binary==2.9.3; python_version >= "3.6"
pycparser==2.21; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pyjwt==2.3.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
python3-openid==3.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pytz==2021.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
requests-oauthlib==1.3.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
requests==2.27.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
six==1.16.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-app-django==4.0.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-core==4.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8"
uritemplate==4.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
urllib3==1.26.8; python_full_version >= "3.6.1" and python_version < "4" and python_full_version < "4.0.0"
whitenoise==6.0.0; python_version >= "3.7"
